Refactor to_decomon() and to_decomon_merge()
- keep a single public function to_decomon
- avoid code duplication
- put the related code in a dedicated module decomon.layers.convert
- instead of globals() use a mapping constructed from vars() applied to
  relevant module. It avoids explicitely import each DecomonLayer and
  make the module imports be used so that it is obvious even for IDE that
  such imports are not to be removed.
