+++ src/decomon/models/convert.py
@@ -9,8 +9,8 @@ from tensorflow.keras.models import Model
 
 from decomon.backward_layers.core import BackwardLayer
 from decomon.backward_layers.crown import Convert2Mode
+from decomon.layers.convert import to_decomon
 from decomon.layers.core import get_mode
-from decomon.layers.decomon_layers import to_decomon
 from decomon.models.backward_cloning import convert_backward
 from decomon.models.forward_cloning import (
     LayerMapDict,
@@ -19,12 +19,7 @@ from decomon.models.forward_cloning import (
     convert_forward_functional_model,
 )
 from decomon.models.models import DecomonModel
-from decomon.models.utils import (
-    ConvertMethod,
-    get_input_tensors,
-    get_input_tensors_keras_only,
-    split_activation,
-)
+from decomon.models.utils import ConvertMethod, get_input_tensors, split_activation
 from decomon.utils import ConvexDomainType, Slope, get_lower, get_upper
