+++ tests/test_conv.py
@@ -7,8 +7,9 @@ import tensorflow.keras.backend as K
 from tensorflow.keras.layers import Conv2D
 from tensorflow.python.keras.backend import _get_available_gpus
 
+from decomon.layers.convert import to_decomon
 from decomon.layers.core import ForwardMode
-from decomon.layers.decomon_layers import DecomonConv2D, to_decomon
+from decomon.layers.decomon_layers import DecomonConv2D
 
 
 def test_Decomon_conv_box(data_format, mode, floatx, helpers):
